+++ packages/api/src/Domain/Client/Auth/AuthApiOperations.ts
@@ -1,0 +1,5 @@
+export enum AuthApiOperations {
+  GenerateRecoveryCodes,
+  GetRecoveryKeyParams,
+  SignInWithRecoveryCodes,
+}