+++ packages/api/src/Domain/Request/Recovery/SignInWithRecoveryCodesRequestParams.ts
@@ -1,0 +1,7 @@
+export interface SignInWithRecoveryCodesRequestParams {
+  api_version: string
+  username: string
+  password: string
+  code_verifier: string
+  recovery_codes: string
+}