+++ packages/ui-services/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.23.5](https://github.com/standardnotes/app/compare/@standardnotes/ui-services@1.23.4...@standardnotes/ui-services@1.23.5) (2023-01-09)
+
+**Note:** Version bump only for package @standardnotes/ui-services
+
+## [1.23.4](https://github.com/standardnotes/app/compare/@standardnotes/ui-services@1.23.3...@standardnotes/ui-services@1.23.4) (2023-01-09)
+
+**Note:** Version bump only for package @standardnotes/ui-services
+
 ## [1.23.3](https://github.com/standardnotes/app/compare/@standardnotes/ui-services@1.23.2...@standardnotes/ui-services@1.23.3) (2023-01-06)
 
 ### Bug Fixes